Code cleanups
git-svn-id: https://svn.atnf.csiro.au/difx@6492 9d4a3555-dc7e-4673-bb7d-d8a0b9a30876

+++ applications/difx2mark4/trunk/src/get_vis.c
@@ -139,11 +139,11 @@ int get_vis (char *vf_name,                   // name of input file
             if (opts->verbose > 1)
                 {
                 if (sizeof(size_t) == 4)
-                    printf ("realloc another mem chunk for visibilities, nvrec %d size %u\n",
+                    printf ("realloc another mem chunk for visibilities, nvrec %d size %lu\n",
                             *nvrec, (size_t) (*nvrec + CHUNK) * (long int) vrsize);
                 else
                     printf ("realloc another Mem chunk for visibilities, nvrec %d size %llu\n",
-                            *nvrec, (size_t) (*nvrec + CHUNK) * (long int) vrsize);
+                            *nvrec, (size_t) (*nvrec + CHUNK) * (long long int) vrsize);
                 }
             *vrec = realloc (*vrec, (size_t) (*nvrec + CHUNK) * (long int) vrsize);
             if (*vrec == NULL)
